vagrant: add scripts to provision vagrant+libvirt and lightos
this commit contains instructions how to setup vagrant and libvirt
provider as an hypervisor in order to run VMs on a single bare-metal
host.

additional instructions are provided to setup network for these VMs,
install ansible and prerequisits for setting up the Ansible-Controller
and example how to install lightos using Ansible on provisioned VMs.