Vagrantfile: don't recration nvme ssds upon vagrant up
lightbits use the emulated ssds as persistent storage that must be
persisted across reboots.

The bug was that each time we called vagrant up the trigger was fired
and recreated the nvme emulated sshs - which delete the lightbits data.

Current fix is to detect if the up trigger is called on already created
VM or not.

if VM is not created it will invoke nvme creation.
If VM already exists - by looking up it's id file - we will skip NVMe
SSDs creation.