lbprox: initial commit of cli that manages proxmox cluster
this cli enable us to connect allocate VMs on a proxmox cluster,
and to define machines that are compatible with lightbits servers
and clients.

it enable us to define emulated SSDs as well as passthrough SSDs
as well as defining a data-network using Proxmox SDN feature.

this CLI supports cloud-init integration and will deploy VMs with
basic cloud-init user-data (will be extended in the future).