lbprox: fix phy ssd attach/detach flow
today proxmox knows automatically to detach the ssd from the kernel
and load the vfio driver and connect it to the VM, but it does not
know how to reconnect it back to the host kernel once it is detached
from the VM and the VM is dead.

This code will detect all detached pci devices and will try to reattach
them to the kernel host.