utils: make the code more robust - detect ACCESS_IP as ip from vmbr0 bridge
till now we treated the LAB_NETWORK to be 192.168.16.0/24
but that is not the case since we would want to run this setup in
different venues not just lightbits lab.

the new logic will detect vmbr0 cidr and will check if the ip is
included in that network it will declare it as access network.