feat: add light_app_path var to lbprox.yaml
till now we counted on WORKSPACE_TOP env var.
in a world outside the lab we don't have WORKSPACE_TOP
so the user must provide this path to the expanded light-app
directory where we can find the ansible playbook and roles.

if not provided we will try to look for WORKSPACE_TOP/light-app
if not provided we will fail with a clear error msg.