feat: add support for uploading tar.gz from url (like cloudsmith)
this feat will allow us to specify external URL of tar.gz
the lbprox will download the file, extract it, look for qcow2
and if present will upload it to the node/nodes specified.

also supported --force flag to create os-images to delete an
existing image and upload a new one.

by default it will not upload if the image is present on one
of the nodes in the cluster.